[webgl] Entry.Variable 분리
- 각 변수타입에 맞춰 클래스 분할
  - 비슷한 코드인데 한 메소드에 분기로 너무 길어서 잘라냈습니다.
- new Entry.Variable -> Entry.Variable.create
  - Entry.Variable 자체를 creator function 으로 수정하려고 했으나,
  instanceOf 비교연산이 꽤 많아서 그대로 두기로 했음
